add share component
