+++ app/layout/community/published-projects/detail/component.tsx
@@ -3,7 +3,6 @@ import React from 'react';
 import { useRouter } from 'next/router';
 
 import { usePublishedProject } from 'hooks/projects';
-import { useToasts } from 'hooks/toast';
 
 import Avatar from 'components/avatar';
 import Backlink from 'layout/statics/backlink';
@@ -11,11 +10,10 @@ import Button from 'components/button';
 import Icon from 'components/icon';
 import Loading from 'components/loading';
 import PublishedProjectMap from 'layout/community/published-projects/detail/map';
+import Share from 'layout/community/published-projects/detail/share';
 import Wrapper from 'layout/wrapper';
 
 import DOWNLOAD_SVG from 'svgs/ui/download.svg?sprite';
-import FACEBOOK_FILLED_SVG from 'svgs/social/facebook-filled.svg?sprite';
-import TWITTER_FILLED_SVG from 'svgs/social/twitter-filled.svg?sprite';
 
 export interface CommunityProjectsDetailProps {
 
@@ -24,7 +22,6 @@ export interface CommunityProjectsDetailProps {
 export const CommunityProjectsDetail: React.FC<CommunityProjectsDetailProps> = () => {
   const { query } = useRouter();
   const { pid } = query;
-  const { addToast } = useToasts();
   const {
     data: publishedProject,
     isFetching: publishedProjectIsFetching,
@@ -34,39 +31,6 @@ export const CommunityProjectsDetail: React.FC<CommunityProjectsDetailProps> = (
   const {
     description, name, planningAreaName, users, scenarios,
   } = publishedProject || {};
-
-  const handleCopy = () => {
-    const copyURL = document.getElementById('copyURLInput').baseURI;
-    navigator.clipboard.writeText(copyURL).then(
-      () => {
-        addToast('success-upload-shapefile', (
-          <>
-            <h2 className="font-medium">Success!</h2>
-            <p className="text-sm">Link copied to clipboard</p>
-          </>
-        ), {
-          level: 'success',
-        });
-      },
-    )
-      .catch(
-        () => {
-          addToast('error-upload-shapefile', (
-            <>
-              <h2 className="font-medium">Error!</h2>
-              <p className="text-sm">
-                The link could not be copied.
-                {' '}
-                <br />
-                Please, try again!
-              </p>
-            </>
-          ), {
-            level: 'error',
-          });
-        },
-      );
-  };
 
   return (
     <Wrapper>
@@ -131,39 +95,7 @@ export const CommunityProjectsDetail: React.FC<CommunityProjectsDetailProps> = (
                   </>
                 )}
               </div>
-              <div>
-                <h3 className="mb-6 text-sm font-semibold text-white">Share</h3>
-                <div className="mb-5">
-                  <button onClick={handleCopy} type="button">
-                    Copy link
-                  </button>
-                  <input className="hidden" type="text" value="" id="copyURLInput" />
-                </div>
-                <div className="mb-5">
-                  <a
-                    className="flex flex-row"
-                    type="button"
-                    role="button"
-                    href={`https://twitter.com/intent/tweet?url=${window.location.toString()}`}
-                    rel="noreferrer"
-                    target="_blank"
-                  >
-                    <Icon icon={TWITTER_FILLED_SVG} className="w-5 h-5 mr-2.5 text-white" />
-                    Twitter
-                  </a>
-                </div>
-                <a
-                  className="flex flex-row"
-                  type="button"
-                  role="button"
-                  href={`https://www.facebook.com/sharer/sharer.php?u=${window.location.toString()}`}
-                  rel="noreferrer"
-                  target="_blank"
-                >
-                  <Icon icon={FACEBOOK_FILLED_SVG} className="w-5 h-5 mr-2.5 text-white" />
-                  Facebook
-                </a>
-              </div>
+              <Share />
             </div>
           </div>
           <div className="w-5/12">
