+++ app/layout/community/published-projects/detail/share/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
